Output less code for opal_engine_check (fixes opal#1965)
The rationale for that is that if/else blocks already share the
closure in JS (we don't output `let` variables at this time yet),
so nothing is expected to break. Another rationale is that old code
produced more code than needed, for example it created an empty
else block even if no else block was present, now the worst case
scenario is adding an empty `nil`.